Resolve missing local properties

+++ resources/application.rb
@@ -28,24 +28,37 @@ def get_down_params
 end
 
 action :up do
+  project_name = new_resource.project_name || current_resource.project_name
+  compose_files = new_resource.compose_files || current_resource.compose_files
+
   execute "running docker-compose up for project #{project_name}" do
     command "docker-compose #{get_compose_params} up #{get_up_params}"
+    environment('PATH' => '/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin')
+    only_if "[ $(docker-compose -f #{compose_files.join(' -f ')} ps -q | wc -l) -eq 0 ]"
     user 'root'
     group 'root'
   end
 end
 
 action :create do
+  project_name = new_resource.project_name || current_resource.project_name
+
   execute "running docker-compose create for project #{project_name}" do
     command "docker-compose #{get_compose_params} create #{get_up_params}"
+    environment('PATH' => '/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin')
     user 'root'
     group 'root'
   end
 end
 
 action :down do
+  project_name = new_resource.project_name || current_resource.project_name
+  compose_files = new_resource.compose_files || current_resource.compose_files
+
   execute "running docker-compose down for project #{project_name}" do
     command "docker-compose #{get_compose_params} down  #{get_down_params}"
+    environment('PATH' => '/usr/local/sbin:/usr/local/bin:/sbin:/bin:/usr/sbin:/usr/bin')
+    not_if "[ $(docker-compose -f #{compose_files.join(' -f ')} ps -q | wc -l) -eq 0 ]"
     user 'root'
     group 'root'
   end
